refactor: move CdkNagSuppressions to src/common
These interfaces describe the shape of cdk-nag's `cdk_nag` resource
metadata. Nothing about them is specific to the SecurityCompliance
aspect, but they lived under src/aspects/security-compliance/types/, so
the ServerAccessLogsBucket construct had to reach into an aspect's
internals to type its own suppressions - the only src/constructs to
src/aspects import in the repo.

Moved verbatim to src/common/cdk-nag.ts, alongside the existing
default-config.ts and validate.ts, and all three importers repointed.
src/common is not re-exported from src/index.ts, so the types stay
internal and out of API.md, same as before the move.
